icell8_stats.py: move 'get_batch_size' function into icell8/utils module.
